requirements.txt
absl-py
colorama
comtypes
duckduckgo_search
keyboard
mediapipe
numpy
opencv-python
playsound
psutil
PyAudio
pyautogui
pycaw
pygetwindow
pyjokes
pytesseract
pyttsx3
requests
screen_brightness_control
SpeechRecognition
Wikipedia-API
